检查完善接口

+++ src/main/java/tech/icoding/sjv/service/ServerInfoService.java
@@ -1,7 +1,9 @@
 package tech.icoding.sjv.service;
 
 import org.springframework.stereotype.Service;
+import tech.icoding.sjv.exception.BadRequestException;
 import tech.icoding.sjv.model.ServerInfo;
+import tech.icoding.sjv.repository.ClientInfoRepository;
 import tech.icoding.sjv.repository.ServerInfoRepository;
 import tech.icoding.sjv.util.CustomSqlServerConnector;
 
@@ -14,21 +16,24 @@ import java.util.List;
 public class ServerInfoService {
 
     private final ServerInfoRepository repository;
+    private final ClientInfoRepository clientInfoRepository;
 
-    public ServerInfoService(ServerInfoRepository repository) {
+    public ServerInfoService(ServerInfoRepository repository, ClientInfoRepository clientInfoRepository) {
         this.repository = repository;
+        this.clientInfoRepository = clientInfoRepository;
     }
 
     public List<ServerInfo> findAll() {
         return repository.findAll();
     }
 
+
     public ServerInfo save(ServerInfo serverInfo) throws Exception {
         boolean validatedConnection = CustomSqlServerConnector.validateConnection(serverInfo.getAddress(),
                 serverInfo.getPort(),
                 serverInfo.getUsername(),
                 serverInfo.getPassword());
-        if(!validatedConnection) throw new Exception("配置信息有误，无法链接数据库");
+        if(!validatedConnection) throw new BadRequestException("配置信息有误，无法链接数据库");
         enableDatabase(serverInfo, "A客户管理");
         return repository.save(serverInfo);
     }
@@ -37,7 +42,15 @@ public class ServerInfoService {
         return repository.findById(id).orElse(null);
     }
 
+    /**
+     * 删除某个服务器信息；删除前需要检查是否有客户使用该服务器，如果有抛出异常
+     * @param id
+     */
     public void delete(Long id) {
+        boolean existsByServerId = clientInfoRepository.existsByServerId(id);
+        if(existsByServerId) {
+            throw new BadRequestException("该服务器有客户使用，无法删除");
+        }
         repository.deleteById(id);
     }
